(BSR)[API] feat: add `AdageMockEndpointHelper` to factorize test_should_raise_404 tests
